Pickle = Trouble.
Moved the Pth and Tf class to config.py because otherwise Pickle cannot serialze the classes for some reason. Pickle is used by the parallel thing
